After check block(from solver2)

+++ app/src/main/java/cdio/group21/litaire/data/Move.kt
@@ -3,6 +3,6 @@ package cdio.group21.litaire.data
 data class Move(
     val isMoveToFoundation: Boolean,
     val card: Card,
-    val indexOfBlock : Int, // source
-    val indexOfDestination : Int
+    val indexOfBlock : Byte, // source
+    val indexOfDestination : Byte
 )
